api_python/requirements.txt
astroid==1.6.0
Django==2.0
django-cors-headers==2.1.0
django-filter==1.1.0
djangorestframework==3.7.3
isort==4.2.15
lazy-object-proxy==1.3.1
Markdown==2.6.9
mccabe==0.6.1
psycopg2==2.7.3.2
ptvsd==3.2.1
pylint==1.8.1
pytz==2017.3
six==1.11.0
wrapt==1.10.11
